+++ astropy/io/fits/util.py
@@ -14,6 +14,8 @@ import tempfile
 import textwrap
 import threading
 import warnings
+import platform
+from distutils.version import LooseVersion
 
 import numpy as np
 
@@ -585,11 +587,32 @@ def fill(text, width, *args, **kwargs):
     return '\n\n'.join(maybe_fill(p) for p in paragraphs)
 
 
+# On MacOS X 10.8 and earlier, there is a bug that causes numpy.fromfile to
+# fail when reading over 2Gb of data. If we detect these versions of MacOS X,
+# we can instead read the data in chunks.
+osx_version = platform.mac_ver()[0]
+if osx_version != '' and LooseVersion(osx_version) < LooseVersion('10.9'):
+    CHUNKED_FROMFILE = True
+else:
+    CHUNKED_FROMFILE = False
+
+
 def _array_from_file(infile, dtype, count, sep):
     """Create a numpy array from a file or a file-like object."""
 
     if isfile(infile):
-        return np.fromfile(infile, dtype=dtype, count=count, sep=sep)
+        if CHUNKED_FROMFILE:
+            chunk_size = int(1024 ** 3 / dtype.itemsize)
+            if count < chunk_size:
+                return np.fromfile(infile, dtype=dtype, count=count, sep=sep)
+            else:
+                array = np.zeros(count, dtype=dtype)
+                for beg in range(0, count, chunk_size):
+                    end = min(count, beg + chunk_size)
+                    array[beg:end] = np.fromfile(infile, dtype=dtype, count=end - beg, sep=sep)
+                return array
+        else:
+            return np.fromfile(infile, dtype=dtype, count=count, sep=sep)
     else:
         # treat as file-like object with "read" method; this includes gzip file
         # objects, because numpy.fromfile just reads the compressed bytes from